requirements.txt
python-binance==1.0.19
pyTelegramBotAPI==4.18.0
pandas>=1.3.0,<2.0.0
numpy>=1.21.0,<1.24.0
matplotlib>=3.4.0,<3.7.0
python-dotenv==1.0.0
requests>=2.28.0
google-generativeai>=0.3.0
flask>=2.3.0
flask-cors>=4.0.0
importlib-metadata>=6.0.0
